+++ app/build.gradle
@@ -260,7 +260,6 @@ dependencies {
     implementation project(':common-utils')
     implementation project(':app-store')
     implementation project(':common-ui')
-    implementation project(':common-ui-experiments')
     internalImplementation project(':common-ui-internal')
     implementation project(':di')
     implementation project(':app-tracking-api')
