+++ app/src/main/java/com/duckduckgo/app/appearance/AppearanceViewModel.kt
@@ -38,7 +38,7 @@ import com.duckduckgo.common.ui.DuckDuckGoTheme.DARK_EXPERIMENT
 import com.duckduckgo.common.ui.DuckDuckGoTheme.LIGHT
 import com.duckduckgo.common.ui.DuckDuckGoTheme.LIGHT_EXPERIMENT
 import com.duckduckgo.common.ui.DuckDuckGoTheme.SYSTEM_DEFAULT
-import com.duckduckgo.common.ui.experiments.BrowserThemingFeature
+import com.duckduckgo.common.ui.internal.experiments.BrowserThemingFeature
 import com.duckduckgo.common.ui.store.ThemingDataStore
 import com.duckduckgo.common.utils.DispatcherProvider
 import com.duckduckgo.di.scopes.ActivityScope
