--- common/common-ui-experiments/build.gradle
@@ -1,56 +1,0 @@
-/*
- * Copyright (c) 2025 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
-    id 'com.squareup.anvil'
-}
-
-apply from: "$rootProject.projectDir/gradle/android-library.gradle"
-
-android {
-    namespace 'com.duckduckgo.common.ui.experiments'
-}
-
-android {
-    anvil {
-        generateDaggerFactories = true // default is false
-    }
-    lintOptions {
-        baseline file("lint-baseline.xml")
-    }
-}
-
-dependencies {
-
-    implementation project(path: ':common-utils')
-    implementation project(path: ':di')
-    anvil project(path: ':anvil-compiler')
-    implementation project(path: ':anvil-annotations')
-    implementation project(path: ':app-build-config-api')
-
-    implementation AndroidX.appCompat
-    implementation Google.android.material
-    implementation AndroidX.constraintLayout
-    implementation AndroidX.core.splashscreen
-    implementation AndroidX.recyclerView
-    implementation AndroidX.lifecycle.viewModelKtx
-    // just to get the dagger annotations
-    implementation Google.dagger
-
-    implementation "androidx.core:core-ktx:_"
-}
