+++ common/common-ui-internal/AndroidManifest.xml
