+++ common/common-ui-internal/build.gradle
@@ -30,13 +30,26 @@ android {
 }
 
 dependencies {
-    implementation project(':common-ui')
-    implementation project(':di')
-    implementation project(':internal-features-api')
-    implementation project(':anvil-annotations')
-
-    // Dagger
-    implementation Google.dagger
 
     implementation "com.squareup.logcat:logcat:_"
+
+    implementation project(path: ':common-utils')
+    implementation project(':common-ui')
+
+    implementation project(path: ':di')
+    anvil project(path: ':anvil-compiler')
+    implementation project(path: ':anvil-annotations')
+    implementation project(path: ':app-build-config-api')
+    implementation project(':internal-features-api')
+    implementation project(path: ':navigation-api')
+
+    implementation AndroidX.appCompat
+    implementation Google.android.material
+    implementation AndroidX.constraintLayout
+    implementation AndroidX.core.splashscreen
+    implementation AndroidX.recyclerView
+    implementation AndroidX.lifecycle.viewModelKtx
+    // just to get the dagger annotations
+    implementation Google.dagger
+
 }
