+++ common/common-ui/build.gradle
@@ -38,7 +38,6 @@ android {
 dependencies {
 
     implementation project(path: ':common-utils')
-    implementation project(path: ':common-ui-experiments')
     implementation project(path: ':di')
     anvil project(path: ':anvil-compiler')
     implementation project(path: ':anvil-annotations')
